Move image loader to top
